Fixed: restored basic functionality
